rename: 사이드바 컴포넌트 이름 변경
